Fix file access after renames
If an NXfield nxdata array has not been read from the file, it is
inaccessible after any NXobject in its path has been renamed because
the original file's path no longer matches the tree. I have created a
new variable, _filepath, to retain the file version until a save. This
is not fully implemented yet. Also, references to the now redundant
treemodel have been removed. It is no longer necessary to inherit the
NXobjects class from the model.