Copy data fixes
Fixes two bugs: 1) If nxdata had not been loaded before copying the
NXfield to a new group, there would be no way to retrieve it from the
file after the copy. The fix will only work if nxdata is smaller than
NX_MEMORY so this will still need to be fixed for larger data. 2) The
signal attribute was erroneously added to the original NXfield when it
was copied to a new NXdata group.