Switch to QDoubleSpinBox
Previously, the x, y, and v axes used a subclass of QSpinBox, which
seemed to have better performance but introduced quirks when trying to
edit the text directly. Now the QDoubleSpinBox performance seems to
have improved. Still using QSpinBox for the z-axis so that the user can
step through the orthogonal direction.