Initial implementation of least-squares fitting
A new Fit Dialog is an option in the Data menu or by right-clicking on
a data item. This allows the addition of several functions, as defined
in the nexpy/api/frills/functions directory. Still in the early stages
of development - parameters can be fixed but not limited or bound yet,
and the results of the fits need to be displayed. However, the results
of a fit can be saved to a scratch NXentry group in w0 on the tree.