Improved saving of data
Now, it is assumed that the save method of all non-NXroot objects
should produce a new file (wrapped in an NXentry group if needed),
i.e., the save method is not used to save modifications to the file of
an object's root group. For that, the save method must be used on the
root group itself.