Fix loading libNeXus.so under Ubuntu 12.04
The libnexus package of Ubuntu (and probably Debian) does not install a symbolic
link under /usr/lib/libNeXus.so to /usr/lib/libNeXus.so.0, hence loading fails.
This is not a fault by the distributions but related to the way loading is done
in Nexpy.

I fixed the problem using find_library of the ctypes.util module, which
automatically searches for a library under all major platforms. I highly suggest
to replace the hand-written code with that, making loading a) cleaner and b)
more reliable.