Refactored tree view
Before this, each NeXus node had an _item attribute linking it to the
tree view item. However, this led to crashes when the tree view item
had been deleted. The new implementation synchronizes the tree view
with the ‘tree’ node whenever a node is changed. This leads to much
simpler code and, so far, more reliable behavior.