Added nxtitle property to NXfield
The Plot Dialog is triggered if there is any error in trying to plot
the node. Since the ‘nxtitle’ property had not been properly defined
for NXfields, an exception was raised even if the field was a valid
signal. Now, it will plot the field if there isn’t another exception.