Fixed root field bug
If a field is added to the file root (in violation of the NeXus
standard), NXFile._writegroup would add the root group name to the root
when writing to the file producing a duplicate tree. Also, refined the
handling of variable length string inputs to use h5py’s check_dtype
function to confirm that the object datatype is a string.