Various enhancements to tree.py
1) Refactored data value reads, with new NXFile._readvalue and
NXFile.open functions, to facilitate subclassing of NXFile for remote
file access. Core memory files are now opened directly in h5py, instead
of through NXFile, so that they are not affected by the lack of
persistence of NXFile file handles.
2) Added a parent attribute to AttrDict to allow attributes to be
written directly to HDF5 files, if the file mode allows it.m Also added
__delitem__ to AttrDict and __delattr__ to NXfield to allow deletion of
attributes.
3) Improved efficiency of h5py file updating. Before this, whole groups
would be rewritten if a single field was added to it.
4) Customized NXlink.update so that links are automatically updated.
Otherwise, NXfield.update would call NXFile._writedata which just
returns the link info without performing an update.