Various changes to updating HDF5 files
Added update and rename functions to NXFile so that the private
functions within NXFile are no longer referenced outside the class.
However, _readvalue was renamed to readvalue so that it can be
referenced. Also added the equivalent writevalue function. These
changes help to encapsulate IO within the NXFile class to make it
easier to subclass in NXFileRemote.