Fixed plotting of individual NXfields
If there were size-1 axes, they were not properly handled in the
PlotDialog. This also revealed a bug in NXfield._setdata (fixed in the
last commit), that meant that, when _uncopied_data was set, it wasn’t
copied to _value if the new field was not stored in a file.