Removed offset in projection panel
The projection panel limits were set by the currently set axis limits,
but then adjusted slightly inwards to make sure the projection
rectangle was visible. However, this also meant the z-axis limits were
different from the limits, which might have been specifically set by
the user. The only drawback is that the rectange won’t be immediately
visible, but it will show up as soon as any limits are changed.