Fix display of magic menu
IPython 3.0.0 has a bug (now fixed in the dev version), in which the
result of ‘lsmagic’ was passed as a dictionary to json.loads, which
expects a string. To ensure backward compatibility with IPython<3.0.0,
we now check whether the result is a dictionary or a string, only
calling json.loads for the latter.