ENH : add qt compatibility
This uses matplotlib's qt compatibility layer to source the Qt classes.

It is a bit odd to use the plotting library to source the Qt layer, but
doing it this way will make sure that the main gui and matplotlib will
stay in sync back-end wise with out having to worry about import orders
and setting mpl rcParams.