Restore multiline editing
The switch to the RichJupyterWidget resulted in losing the automatic
detection of incomplete code typed in the shell. This is apparently due
to limitations in the current in-process kernel code in the qtconsole
and ipykernel packages, part of the Jupyter distribution. Until it is
fixed, this is a work-around that restores multiline editing.