Turn of matplotlib.use warning
The Matplotlib use function should only be used once and, by default,
issues a warning if there is a second call. This causes problems within
the nexusformat package, which checks to see if the plot call is from
within NeXpy, i.e., to allow plots from the IPython shell. However,
within the IPython Qt console application, this check triggers a second
check, so it is necessary to suppress warnings.