Enhance options for resetting limits
The default axis or signal limits, which can be changed through the
sliders or text boxes, are reset by clicking on the Home button of the
Options tab or invoking the “Reset Plot Limits” menu item. Previously,
right-clicking within the plot had the same effect, but this has been
changed to reset the axes but not the signal limits. This is often the
preferred behavior, so this gives a quick method of enabling it.