Update values for lmfit version > 0.8.3
The latest version of lmfit does not update the input parameters in
place, but puts the fit results into a new MinmizerResult object. This
copies them back again, since that is how we have implemented lmfit in
NeXpy. Although this seems to be the right thing to do, my initial
tests do not show any change in the MinimizerResult parameters, so
there may be other problems to fix.