Transfer parameters in the residuals function
The new parameter values were not being passed to the model functions
because the old version assumed that the underlying Parameter objects
were the same, but with new values. In the new version of lmfit, that
is no longer true - the Parameter objects used by the model have to be
explicitly updated.