Add astropy convolution as a smoothing option
This module handles NaNs properly, so it provides a more satisfactory
smoothing when there are gaps in the data. The smoothed data is still
pixellated, so the matplotlib bicubic smoothing is applied if imshow is
called. The ‘convolve’ option is only added to the interpolation list
if the astropy module is installed.