Allow symmetric signal plots
If a divergent color map is chosen, the minimum signal value is fixed
to be the negative of the maximum signal. The minimum double-spin box
and sliders are disabled. Dividers are added to the color map combobox
to make clear which maps are divergent. Selecting the log checkbox
implements a symmetric log scale.