Fix plotting with changed axes
When calling replot_data with newaxis=True, the x-axis and y-axis
minimum and maximum values may not correspond to the minimum and
maximum values of the data since they can be changed by the user. This
leads to incorrect ‘extent’ values when calling imshow. This uses the
minimum and maximum values of the bin boundaries.