Change default slice in >2D plots
This used to select the slices with index 0, but this was almost never
what was desired (in our experience). For four-dimensional data, this
required changing all the z-axis values before a sensible plot was
displayed. Any choice is arbitrary, but selecting a z-value of 0.0
works more than most. If the z-array does not include 0, then this
defaults to the 0 index as before.