Add resize function to NXFigureManager
When saving a figure, Matplotlib’s FigureManagerQT class performs a
redundant resize of the manager window, with a fixed height increment
to account for the standard Matplotlib tool and status bars. It is
necessary for NeXpy to replace the resize function to cope with the
different GUI layout. The NXPlotView minimum width is also now 24
pixels later than the canvas minimum width to allow for the window
margin.