Add a replacement for sys.excepthook
This was required because PyQt 5.5 now terminates the program whenever it encounters an exception, even when it is caught, unless a cusotomized  exception hook is defined. The new report_exception function now catches all uncaught exceptions, apart from those handled by the Jupyter shell. The exception is logged and displayed in a dialog box with its traceback.