Enable plotting of a minor grid on skewed axes
This is not currently supported in Matplotlib, although it does calculate the minor tick locations. This draws or removes both the major and minor grid locations manually for skewed axes. When resetting the limits to the original (i.e., calling the home function), the grid is redrawn because it is not reset by Matplotlib.