Reconfigure the MainWindow close
Previously, the QMainwindow close function was customized to confirm requests to quit NeXpy. However, the recommended behavior is to put such confirmation requests in a customized closeEvent dialog, which is triggered by calls to the QMainwindow close function. This may fix issues on some systems where NeXpy does not shut down properly.