Prevent double triggers in spin boxes
Holding the up or down arrows of both NXSpinBox and NXDoubleSpinBox instances can cause a second stepBy call if processing the step takes too much time. This is interpreted as the user holding down the arrow for multiple steps. The solution is to override timerEvent to check the current state of the mouse before triggering another step.