Remover redundant axis checks
The read_minbox and read_maxbox functions are triggered by valueChanged signals. The only way for the minbox and maxbox values to be out of order is if the user edited the box. This is handled by edit_minbox and edit_maxbox, which are triggered by editingFinished signals. The call to setValue here causes issues because it resets the old_value variable and prevents edit_minbox and edit_maxbox from doing anything.